feat: add styled components
This commit add the styled components in the project, allowing us
to manipulate our css using javascript in the separate files.

We created the @types folder where we are overlaping the default
theme for styled components, and use our global default theme.

Doing a test using styled components in the Button component.